# memstore: optimize