--- a/graph/memstore/iterator.go
+++ b/graph/memstore/iterator.go
@@ -120,22 +120,23 @@ func (it *Iterator) checkValid(index int64) bool {
 
 func (it *Iterator) Next() bool {
 	graph.NextLogIn(it)
-
 	if it.iter == nil {
 		return graph.NextLogOut(it, false)
 	}
-	result, _, err := it.iter.Next()
-	if err != nil {
-		if err != io.EOF {
-			it.err = err
+	for {
+		result, _, err := it.iter.Next()
+		if err != nil {
+			if err != io.EOF {
+				it.err = err
+			}
+			return graph.NextLogOut(it, false)
 		}
-		return graph.NextLogOut(it, false)
+		if !it.checkValid(result) {
+			continue
+		}
+		it.result = result
+		return graph.NextLogOut(it, true)
 	}
-	if !it.checkValid(result) {
-		return it.Next()
-	}
-	it.result = result
-	return graph.NextLogOut(it, true)
 }
 
 func (it *Iterator) Err() error {