```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkNamePredicate-8                      197757        204327        +3.32%
BenchmarkLargeSetsNoIntersection-8            4130486       3369865       -18.41%
BenchmarkVeryLargeSetsSmallIntersection-8     7823824       6172993       -21.10%
BenchmarkHelplessContainsChecker-8            2606138       2199486       -15.60%
BenchmarkHelplessNotContainsFilms-8           762388        693591        -9.02%
BenchmarkHelplessNotContainsActors-8          11091629      7148805       -35.55%
BenchmarkNetAndSpeed-8                        878364        794230        -9.58%
BenchmarkKeanuAndNet-8                        605243        598530        -1.11%
BenchmarkKeanuAndSpeed-8                      722900        689371        -4.64%
BenchmarkKeanuOther-8                         2755144       2276770       -17.36%
BenchmarkKeanuBullockOther-8                  4661789       3791485       -18.67%
BenchmarkSaveBogartPerformances-8             332514        312122        -6.13%

benchmark                                     old allocs     new allocs     delta
BenchmarkNamePredicate-8                      427            426            -0.23%
BenchmarkLargeSetsNoIntersection-8            12453          11050          -11.27%
BenchmarkVeryLargeSetsSmallIntersection-8     28221          24190          -14.28%
BenchmarkHelplessContainsChecker-8            10937          9786           -10.52%
BenchmarkHelplessNotContainsFilms-8           3070           2894           -5.73%
BenchmarkHelplessNotContainsActors-8          51836          37899          -26.89%
BenchmarkNetAndSpeed-8                        4368           4205           -3.73%
BenchmarkKeanuAndNet-8                        3176           3139           -1.16%
BenchmarkKeanuAndSpeed-8                      3664           3536           -3.49%
BenchmarkKeanuOther-8                         12833          11128          -13.29%
BenchmarkKeanuBullockOther-8                  20528          17708          -13.74%
BenchmarkSaveBogartPerformances-8             1148           1099           -4.27%

benchmark                                     old bytes     new bytes     delta
BenchmarkNamePredicate-8                      26373         26337         -0.14%
BenchmarkLargeSetsNoIntersection-8            535505        504604        -5.77%
BenchmarkVeryLargeSetsSmallIntersection-8     1287758       1210889       -5.97%
BenchmarkHelplessContainsChecker-8            773145        744658        -3.68%
BenchmarkHelplessNotContainsFilms-8           174363        169942        -2.54%
BenchmarkHelplessNotContainsActors-8          1597660       1284654       -19.59%
BenchmarkNetAndSpeed-8                        278525        275868        -0.95%
BenchmarkKeanuAndNet-8                        203251        202581        -0.33%
BenchmarkKeanuAndSpeed-8                      223859        221799        -0.92%
BenchmarkKeanuOther-8                         663184        629927        -5.01%
BenchmarkKeanuBullockOther-8                  987280        932568        -5.54%
BenchmarkSaveBogartPerformances-8             65424         64432         -1.52%
```